# btf: support marshaling Type