--- a/internal/btf/types_test.go
+++ b/internal/btf/types_test.go
@@ -167,7 +167,7 @@ func newCyclicalType(n int) Type {
 		case 3:
 			prev = &Typedef{Type: prev}
 		case 4:
-			prev = &Array{Type: prev}
+			prev = &Array{Type: prev, Index: prev}
 		}
 	}
 	ptr.Target = prev