… internal/btf/marshal_test.go
… internal/btf/strings_test.go
… internal/btf/traversal_test.go