This is a first cut at adding a `btf.Builder` which allows turning `Type` into BTF that the kernel understands. The code can round-trip vmlinux BTF and then load the result into the kernel, which is a good indication that we covered most of the bases.

As it stands, the builder is not exported or used in the package, since there is currently no way to add strings necessary for `LineInfo` to the BTF. This should become a lot easier once #606 lands.

Best reviewed per commit.

btf: add btfInt
    
    Replace intEncoding with a type. This allows adding setters.

btf: add Array.Index
    
    The kernel refuses to load an array type without an index type, so probably
    btf.rst is outdated.

btf: move typeDeque to separate file
    
    Split typeDeque into a separate file to reduce the amount of code in types.go.

btf: add stringTableBuilder
    
    A stringTableBuilder helps build a BTF-style string table, which delimits
    individual strings with a 0 byte. Each unique string is only added to the
    table once.

btf: add builder
    
    Allow building BTF wire format from scratch. We can compare a full vmlinux build to
    just re-encoding raw type data as we currently do in Spec.marshal:
    
        name                    time/op
        BuildVmlinux/builder-4  81.7ms ± 1%
        BuildVmlinux/native-4   23.5ms ± 2%
    
        name                    alloc/op
        BuildVmlinux/builder-4  46.5MB ± 0%
        BuildVmlinux/native-4   20.1MB ± 0%
    
        name                    allocs/op
        BuildVmlinux/builder-4    700k ± 0%
        BuildVmlinux/native-4     394k ± 0%
    
    A full rebuild is about 4x slower and uses 2x the memory. We can't use the builder
    to replace Spec.marshal yet, since we don't have a good way to include LineInfos
    in the generated BTF. This will become easier once LineInfos are stored in
    Instruction metadata.
    
    Updates #616

btf: replace Type.walk with walkType
    
    Type.walk takes a *typeDeque as argument. Go escape analysis doesn't see through
    interface boundaries and therefore always allocates such pointers on the heap.
    This forces typeDeque to always be heap allocated, which makes flattenType more
    expensive than it has to be.
    
    Replace the interface function with walkType, which essentially contains a big
    type switch. This improves the BTF builder benchmark:
    
        name                    old time/op    new time/op    delta
        BuildVmlinux/builder-4    81.7ms ± 1%    79.2ms ± 1%   -3.09%  (p=0.029 n=4+4)
        BuildVmlinux/native-4     23.5ms ± 2%    23.3ms ± 2%     ~     (p=0.200 n=4+4)
    
        name                    old alloc/op   new alloc/op   delta
        BuildVmlinux/builder-4    46.5MB ± 0%    40.6MB ± 0%  -12.75%  (p=0.029 n=4+4)
        BuildVmlinux/native-4     20.1MB ± 0%    20.1MB ± 0%     ~     (p=0.800 n=4+4)
    
        name                    old allocs/op  new allocs/op  delta
        BuildVmlinux/builder-4      700k ± 0%      576k ± 0%  -17.64%  (p=0.029 n=4+4)
        BuildVmlinux/native-4       394k ± 0%      394k ± 0%     ~     (all equal)